Refactor files to setup BLoC pattern
