+++ movie_shape/lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:movie_shape/helpers/constants.dart';
 import 'package:movie_shape/views/widgets/nav_rail.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 void main() {
   runApp(const MainApp());
@@ -13,15 +12,9 @@ class MainApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     var materialApp = MaterialApp(
-        theme: ThemeData(scaffoldBackgroundColor: AppConstants.primaryColour),
-        home: NavRail());
-    setupEmptyList();
+      theme: ThemeData(scaffoldBackgroundColor: AppConstants.primaryColour),
+      home: NavRail()
+    );
     return materialApp;
   }
-}
-
-Future<void> setupEmptyList() async {
-  final prefs = await SharedPreferences.getInstance();
-  await prefs.setStringList("favourites", []);
-  print("empty list set up");
 }
