+++ movie_shape/lib/views/watchlist/state/watchlist_bloc/watchlist_bloc.dart
@@ -7,7 +7,6 @@ part 'watchlist_state.dart';
 
 class WatchlistBloc extends Bloc<WatchlistEvent, WatchlistState> {
   WatchlistBloc() : super(WatchlistInitial()) {
-    //
     on<LoadWatchlist>((event, emit) async {
       try {
         emit(WatchlistLoading());
