+++ movie_shape/lib/views/widgets/nav_rail.dart
@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:movie_shape/helpers/constants.dart';
-import 'package:movie_shape/views/pages/favourites_page.dart';
 import 'package:movie_shape/views/pages/home_page.dart';
-import 'package:movie_shape/views/pages/film_detail_page.dart';
-import 'package:movie_shape/models/film.dart';
-import 'package:movie_shape/repository/film_repo_implemented.dart';
+import 'package:movie_shape/views/favourites/favourites_page.dart';
 import 'package:movie_shape/views/watchlist/watchlist_page.dart';
 
 class NavRail extends StatefulWidget {
